Make OpenMP support in built darknet configurable
This allows users to disable OpenMP support in darknet even if that is detected, or assert that
OpenMP support is enabled.

OpenMP support in darknet may not be desirable when one uses GPU, wants to reduce the number of
linked libraries etc, wants to limit CPU parallelism...

Makefile build of darknet allows one to pass `OPENMP`, cmake build doesn't have an explicit flag
for it. But it uses `find_packate(OpenMP)` which is controllable in a generic way.

An alternative to introducing an env variable would be to introduce a cargo feature flag, but that
wouldn't be able to nicely represent the autodetection.